requirements.txt
ibm_db
polars
pyyaml
python-dotenv
kubernetes